# Precompute the normalized type names of DataTypes and use string.Create on NS2.1 for faster normalization.