--- a/src/csharp/Microsoft.Spark/Sql/Types/DataType.cs
+++ b/src/csharp/Microsoft.Spark/Sql/Types/DataType.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Linq;
+using System.Diagnostics;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using Microsoft.Spark.Interop.Ipc;
@@ -39,10 +39,13 @@ namespace Microsoft.Spark.Sql.Types
             typeof(MapType),
             typeof(StructType) };
 
+        private static string[] s_simpleTypeNormalizedNames = null;
+        private static string[] s_complexTypeNormalizedNames = null;
+
         /// <summary>
         /// Normalized type name.
         /// </summary>
-        public string TypeName => NormalizeTypeName(GetType().Name);
+        public string TypeName => NormalizeTypeName(GetType());
 
         /// <summary>
         /// Simple string version of the current data type.
@@ -134,19 +137,20 @@ namespace Microsoft.Spark.Sql.Types
                 var typeJObject = (JObject)json;
                 if (typeJObject.TryGetValue("type", out JToken type))
                 {
-                    Type complexType = s_complexTypes.FirstOrDefault(
-                        (t) => NormalizeTypeName(t.Name) == type.ToString());
+                    string typeName = type.ToString();
 
-                    if (complexType != default)
+                    int typeIndex = ComplexTypeIndex(typeName);
+
+                    if (typeIndex != -1)
                     {
                         return (DataType)Activator.CreateInstance(
-                            complexType,
+                            s_simpleTypes[typeIndex],
                             BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance,
                             null,
                             new object[] { typeJObject },
                             null);
                     }
-                    else if (type.ToString() == "udt")
+                    else if (typeName == "udt")
                     {
                         throw new NotImplementedException();
                     }
@@ -169,12 +173,12 @@ namespace Microsoft.Spark.Sql.Types
         private static DataType ParseSimpleType(JToken json)
         {
             string typeName = json.ToString();
-            Type simpleType = s_simpleTypes.FirstOrDefault(
-                (t) => NormalizeTypeName(t.Name) == typeName);
 
-            if (simpleType != default)
+            int typeIndex = SimpleTypeIndex(typeName);
+
+            if (typeIndex != -1)
             {
-                return (DataType)Activator.CreateInstance(simpleType);
+                return (DataType)Activator.CreateInstance(s_simpleTypes[typeIndex]);
             }
 
             Match decimalMatch = DecimalType.s_fixedDecimal.Match(typeName);
@@ -191,9 +195,99 @@ namespace Microsoft.Spark.Sql.Types
         /// <summary>
         /// Remove "Type" from the end of type name and lower cases to align with Scala type name.
         /// </summary>
+        /// <param name="type">The type to normalize.</param>
+        /// <returns>Normalized type name.</returns>
+        private static string NormalizeTypeName(Type type)
+        {
+            if (s_simpleTypeNormalizedNames == null)
+            {
+                Debug.Assert(s_complexTypeNormalizedNames == null);
+                BuildNormalizedStringMapping();
+            }
+
+            for (int i = 0; i < s_simpleTypes.Length; i++)
+            {
+                if (s_simpleTypes[i] == type)
+                {
+                    return s_simpleTypeNormalizedNames[i];
+                }
+            }
+
+            for (int i = 0; i < s_complexTypes.Length; i++)
+            {
+                if (s_complexTypes[i] == type)
+                {
+                    return s_complexTypeNormalizedNames[i];
+                }
+            }
+
+            return NormalizeTypeName(type.Name);
+        }
+
+        /// <summary>
+        /// Remove "Type" from the end of type name and lower cases to align with Scala type name.
+        /// </summary>
         /// <param name="typeName">Type name to normalize</param>
         /// <returns>Normalized type name</returns>
-        private static string NormalizeTypeName(string typeName) =>
-            typeName.Substring(0, typeName.Length - 4).ToLower();
+        private static string NormalizeTypeName(string typeName)
+        {
+#if NETSTANDARD2_1
+            return string.Create(typeName.Length - 4, typeName, (span, typeName) =>
+            {
+                typeName.AsSpan(0, typeName.Length - 4).ToLower(span, CultureInfo.CurrentCulture);
+            });
+#else
+            return typeName.Substring(0, typeName.Length - 4).ToLower();
+#endif
+        }
+
+        /// <summary>
+        /// Uses the built up normalized type name cache to find the index of the simple type that matches the passed in <paramref name="typeName"/>.
+        /// </summary>
+        /// <param name="typeName">Normalized simple type name to compare against.</param>
+        /// <returns>The index of the simple type within the list that matches the type name, if found. Otherwise, -1.</returns>
+        private static int SimpleTypeIndex(string typeName)
+        {
+            if (s_simpleTypeNormalizedNames == null)
+            {
+                Debug.Assert(s_complexTypeNormalizedNames == null);
+                BuildNormalizedStringMapping();
+            }
+            return s_simpleTypeNormalizedNames.AsSpan().IndexOf(typeName);
+        }
+
+        /// <summary>
+        /// Uses the built up normalized type name cache to find the index of the complex type that matches the passed in <paramref name="typeName"/>.
+        /// </summary>
+        /// <param name="typeName">Normalized complex type name to compare against.</param>
+        /// <returns>The index of the complex type within the list that matches the type name, if found. Otherwise, -1.</returns>
+        private static int ComplexTypeIndex(string typeName)
+        {
+            if (s_simpleTypeNormalizedNames == null)
+            {
+                Debug.Assert(s_complexTypeNormalizedNames == null);
+                BuildNormalizedStringMapping();
+            }
+            return s_complexTypeNormalizedNames.AsSpan().IndexOf(typeName);
+        }
+
+        /// <summary>
+        /// Builds up the normalized type name cache for both simple and complex types which is used for faster type name look up.
+        /// </summary>
+        private static void BuildNormalizedStringMapping()
+        {
+            s_simpleTypeNormalizedNames = new string[s_simpleTypes.Length];
+            s_complexTypeNormalizedNames = new string[s_complexTypes.Length];
+
+            for (int i = 0; i < s_simpleTypes.Length; i++)
+            {
+                s_simpleTypeNormalizedNames[i] = NormalizeTypeName(s_simpleTypes[i].Name);
+            }
+
+            for (int i = 0; i < s_complexTypes.Length; i++)
+            {
+                s_complexTypeNormalizedNames[i] = NormalizeTypeName(s_complexTypes[i].Name);
+            }
+        }
     }
 }