--- a/src/csharp/Microsoft.Spark/Sql/Types/DataType.cs
+++ b/src/csharp/Microsoft.Spark/Sql/Types/DataType.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Diagnostics;
 using System.Globalization;
+using System.Linq;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using Microsoft.Spark.Interop.Ipc;
@@ -40,13 +41,20 @@ namespace Microsoft.Spark.Sql.Types
             typeof(MapType),
             typeof(StructType) };
 
-        private static string[] s_simpleTypeNormalizedNames = null;
-        private static string[] s_complexTypeNormalizedNames = null;
+        private static readonly Lazy<string[]> s_simpleTypeNormalizedNames =
+            new Lazy<string[]>(
+                () => s_simpleTypes.Select(t => NormalizeTypeName(t.Name)).ToArray());
+
+        private static readonly Lazy<string[]> s_complexTypeNormalizedNames =
+            new Lazy<string[]>(
+                () => s_complexTypes.Select(t => NormalizeTypeName(t.Name)).ToArray());
+
+        private string _typeName;
 
         /// <summary>
         /// Normalized type name.
         /// </summary>
-        public string TypeName => NormalizeTypeName(GetType());
+        public string TypeName => _typeName ?? (_typeName = NormalizeTypeName(GetType().Name));
 
         /// <summary>
         /// Simple string version of the current data type.
@@ -140,7 +148,7 @@ namespace Microsoft.Spark.Sql.Types
                 {
                     string typeName = type.ToString();
 
-                    int typeIndex = ComplexTypeIndex(typeName);
+                    int typeIndex = Array.IndexOf(s_complexTypeNormalizedNames.Value, typeName);
 
                     if (typeIndex != -1)
                     {
@@ -175,7 +183,7 @@ namespace Microsoft.Spark.Sql.Types
         {
             string typeName = json.ToString();
 
-            int typeIndex = SimpleTypeIndex(typeName);
+            int typeIndex = Array.IndexOf(s_simpleTypeNormalizedNames.Value, typeName);
 
             if (typeIndex != -1)
             {
@@ -196,38 +204,6 @@ namespace Microsoft.Spark.Sql.Types
         /// <summary>
         /// Remove "Type" from the end of type name and lower cases to align with Scala type name.
         /// </summary>
-        /// <param name="type">The type to normalize.</param>
-        /// <returns>Normalized type name.</returns>
-        private static string NormalizeTypeName(Type type)
-        {
-            if (s_simpleTypeNormalizedNames == null)
-            {
-                Debug.Assert(s_complexTypeNormalizedNames == null);
-                BuildNormalizedStringMapping();
-            }
-
-            for (int i = 0; i < s_simpleTypes.Length; i++)
-            {
-                if (s_simpleTypes[i] == type)
-                {
-                    return s_simpleTypeNormalizedNames[i];
-                }
-            }
-
-            for (int i = 0; i < s_complexTypes.Length; i++)
-            {
-                if (s_complexTypes[i] == type)
-                {
-                    return s_complexTypeNormalizedNames[i];
-                }
-            }
-
-            return NormalizeTypeName(type.Name);
-        }
-
-        /// <summary>
-        /// Remove "Type" from the end of type name and lower cases to align with Scala type name.
-        /// </summary>
         /// <param name="typeName">Type name to normalize</param>
         /// <returns>Normalized type name</returns>
         private static string NormalizeTypeName(string typeName)
@@ -240,55 +216,6 @@ namespace Microsoft.Spark.Sql.Types
 #else
             return typeName.Substring(0, typeName.Length - 4).ToLower();
 #endif
-        }
-
-        /// <summary>
-        /// Uses the built up normalized type name cache to find the index of the simple type that matches the passed in <paramref name="typeName"/>.
-        /// </summary>
-        /// <param name="typeName">Normalized simple type name to compare against.</param>
-        /// <returns>The index of the simple type within the list that matches the type name, if found. Otherwise, -1.</returns>
-        private static int SimpleTypeIndex(string typeName)
-        {
-            if (s_simpleTypeNormalizedNames == null)
-            {
-                Debug.Assert(s_complexTypeNormalizedNames == null);
-                BuildNormalizedStringMapping();
-            }
-            return s_simpleTypeNormalizedNames.AsSpan().IndexOf(typeName);
-        }
-
-        /// <summary>
-        /// Uses the built up normalized type name cache to find the index of the complex type that matches the passed in <paramref name="typeName"/>.
-        /// </summary>
-        /// <param name="typeName">Normalized complex type name to compare against.</param>
-        /// <returns>The index of the complex type within the list that matches the type name, if found. Otherwise, -1.</returns>
-        private static int ComplexTypeIndex(string typeName)
-        {
-            if (s_simpleTypeNormalizedNames == null)
-            {
-                Debug.Assert(s_complexTypeNormalizedNames == null);
-                BuildNormalizedStringMapping();
-            }
-            return s_complexTypeNormalizedNames.AsSpan().IndexOf(typeName);
-        }
-
-        /// <summary>
-        /// Builds up the normalized type name cache for both simple and complex types which is used for faster type name look up.
-        /// </summary>
-        private static void BuildNormalizedStringMapping()
-        {
-            s_simpleTypeNormalizedNames = new string[s_simpleTypes.Length];
-            s_complexTypeNormalizedNames = new string[s_complexTypes.Length];
-
-            for (int i = 0; i < s_simpleTypes.Length; i++)
-            {
-                s_simpleTypeNormalizedNames[i] = NormalizeTypeName(s_simpleTypes[i].Name);
-            }
-
-            for (int i = 0; i < s_complexTypes.Length; i++)
-            {
-                s_complexTypeNormalizedNames[i] = NormalizeTypeName(s_complexTypes[i].Name);
-            }
         }
     }
 }