I noticed some areas where the implementation could be made more performant by avoiding using Linq and caching the normalized type names for the simple and complex DataTypes. That way, they don't have to be computed every time. I am not sure if the code paths are perf critical, in which case, I would understand if this PR isn't useful. If that's the case, please let me know.

Also, added a call to use `string.Create` to make normalizing the type names a bit faster (and allocate less). If adding a netstandard2.1 TFM isn't feasible, please let me know and I can revert that part of the change.

The behavior isn't being changed at all, hence existing tests should be sufficient.

``` ini

BenchmarkDotNet=v0.11.5.1159-nightly, OS=Windows 10.0.18362
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.100-preview1-014459
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT
  Job-PDGMZC : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  IterationTime=250.0000 ms  MaxIterationCount=10  
MinIterationCount=5  WarmupCount=1  

```
Accessing the `TypeName` property for the known simple/complex types got much faster:

|                Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|-------:|------:|------:|----------:|
| TypeName_old | 41.71 ns | 0.684 ns | 0.304 ns | 41.61 ns | 41.31 ns | 42.22 ns |  1.00 | 0.0095 |     - |     - |      40 B |
|     TypeName_new | 14.52 ns | 0.360 ns | 0.188 ns | 14.53 ns | 14.25 ns | 14.72 ns |  0.35 |      - |     - |     - |         - |

Normalizing a new type name (using string.Create) reduces allocations:

|                Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|-------:|------:|------:|----------:|
| NormalizeTypeName_old | 42.60 ns | 1.359 ns | 0.899 ns | 42.54 ns | 41.42 ns | 43.86 ns |  1.00 | 0.0190 |     - |     - |      80 B |
|     NormalizeTypeName_new | 38.42 ns | 0.762 ns | 0.453 ns | 38.22 ns | 38.02 ns | 39.35 ns |  0.90 |      - |     - |     - |    40 B |

cc @eerhardt