@ahsonkhan 2019-12-07
Fyi, the `ParseSimpleType` and `ParseDataType` methods got faster and allocate less (when the type was found in the cache):

|    Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| ParseSimpleType_old | 465.22 ns | 35.651 ns | 23.581 ns | 461.61 ns | 429.68 ns | 507.12 ns |  1.00 | 0.0875 |     - |     - |     368 B |
| ParseSimpleType_new |  88.88 ns |  7.526 ns |  4.978 ns |  87.10 ns |  84.02 ns |  96.57 ns |  0.19 | 0.0057 |     - |     - |      24 B |

I suspect `ParseDataType` will get even faster once this change is in (and we avoid using `JToken.Parse`): https://github.com/dotnet/spark/pull/358