@eerhardt 2019-12-09
> Do you see any cons of targeting multiple frameworks going forward?

The only real con is that you need to build multiple times, but since this repo is small, it won't matter. 

This change will only help future code that wants to take advantage of new APIs (typically for perf reasons, like this change).